+++ avocado/core/restclient/cli/app.py
@@ -23,9 +23,9 @@ import importlib
 import functools
 
 from avocado import settings
-from avocado.restclient import connection
 from avocado.core import output
 from avocado.core import exit_codes
+from avocado.core.restclient import connection
 from avocado.core.restclient.cli import parser
 
 __all__ = ['App']
