--- examples/api/restclient/request.py
@@ -1,7 +1,0 @@
-# Copyright: Red Hat Inc. 2015
-# Author: Cleber Rosa <cleber@redhat.com>
-
-from avocado.restclient import connection
-
-c = connection.get_default()
-print(c.request("version"))
