+++ requirements.txt
@@ -9,5 +9,5 @@ libvirt-python>=1.2.9
 pyliblzma>=0.5.3
 # HTML report plugin (avocado.core.plugins.htmlresult)
 pystache>=0.5.3
-# REST client (avocado.restclient)
+# REST client (avocado.core.restclient)
 requests>=1.2.3
